streaming/requirements.txt
kafka-python-ng==2.2.3
requests==2.31.0
python-dotenv==1.0.0
alpha-vantage==2.3.1
finnhub-python==2.4.25
praw==7.7.1
newsapi-python==0.2.7
pandas==2.1.3
apscheduler==3.10.4
vaderSentiment==3.3.2
beautifulsoup4==4.12.2
lxml
tweepy==4.14.0
httpx
aiohttp
textblob
pandoc
ddgs>=9.7.0
newspaper3k==0.2.8
lxml_html_clean